Update Search.py
Removed the replace and replace all functionality because it doesn't work to select all the effects at once in separate layers. Additionally the find functionality was used more than the replace functionality so I removed it entirely.

Secondly added the layers and effects button to isolate whether the user wants to search layers or they want to search effects.