Update Search.py
Added an event dispatcher to get the latest layer stack changes. Unfortunately I got a few glitches where the layers selected are locked to the selected nodes, current workaround is using a boolean to switch between manual selection and the selection from the find functionality. Going to try to look for alternative solution.