Update Search.py
Did some clean up and added a function that will get the latest changes in the layer stack and will update the root layers. I also adjusted the handle_text_change and navigation functions so that they will get the latest matched items and will iterate over the new layers in the matched items.