Update Search.py
Adding a function that is printing out all the matched items. Some changes is that I grouped find layers and find effects function together, and then I separated matched items to matched layers and matched effects so that the switch view will isolate the matched items based on the view.